backend/researcher/pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "Alex Researcher Agent (GCP)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core API
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    # HTTP + utils
    "httpx>=0.28.1",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
    # Agent framework
    "openai-agents[litellm]>=0.2.4",
    # Browser automation (MCP)
    "playwright>=1.54.0",
    # Config / validation
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    # ✅ GCP (REQUIRED)
    "google-auth>=2.35.0",
    "google-cloud-aiplatform>=1.70.0",
    "sqlalchemy>=2.0.49",
    "psycopg2-binary>=2.9.11",
]
